+++ src/components/GanttView.vue
@@ -13,6 +13,7 @@ import GanttMilestone from './GanttMilestone.vue'
 import GanttNonWorking from './GanttNonWorking.vue'
 import GanttPeriods from './GanttPeriods.vue'
 import GanttSlack from './GanttSlack.vue'
+import GanttSummaryBar from './GanttSummaryBar.vue'
 import GanttTask from './GanttTask.vue'
 import GanttTaskList from './GanttTaskList.vue'
 import GanttTimeline from './GanttTimeline.vue'
@@ -119,6 +120,9 @@ const scrollStyle = computed(() => {
             <template v-if="$slots.row" #row="rowProps">
               <slot name="row" v-bind="rowProps" />
             </template>
+            <template v-if="$slots['row-suffix']" #row-suffix="suffixProps">
+              <slot name="row-suffix" v-bind="suffixProps" />
+            </template>
             <template v-if="$slots.rowEditor" #rowEditor="editorProps">
               <slot name="rowEditor" v-bind="editorProps" />
             </template>
@@ -148,6 +152,14 @@ const scrollStyle = computed(() => {
               <slot name="groupBar" v-bind="groupBarProps" />
             </template>
           </GanttGroupBar>
+        </slot>
+
+        <slot name="summary-bars" :rows="visibleRows">
+          <GanttSummaryBar>
+            <template v-if="$slots.summaryBar" #default="summaryBarProps">
+              <slot name="summaryBar" v-bind="summaryBarProps" />
+            </template>
+          </GanttSummaryBar>
         </slot>
 
         <slot name="baselines" :tasks="visibleTasks">